Move phonetype enum class into separate enum folder
